pyproject.toml
[build-system]
requires = ["setuptools>=69.0.3", "wheel>=0.42.0"]
build-backend = "setuptools.build_meta"

[project]
name = "tcs-app"
version = "1.0.0"
description = "TCS Application"
requires-python = ">=3.11"
dependencies = [
    "Flask==2.3.3",
    "pandas==2.0.3",
    "numpy==1.24.3",
    "openpyxl==3.1.2",
    "python-dotenv==1.0.0",
    "Werkzeug==2.3.7",
    "Pillow==10.2.0",
    "gunicorn==21.2.0",
]

[tool.setuptools]
packages = ["tcs_app"] 